+++ ChangeLog.md
@@ -1,12 +1,12 @@
 # Changelog
 
 ## v1.0.0
-- All Redis 3.0 commands are supported
+- Major commands under Redis 3.0 are supported
 
 ## v0.9.x (2022-xx-xx)
 ###  Supported
 - MSET, MSETNX, MGET
-- HSET, HGET, HMSET, HMGET
+- HSET, HGET, HGETALL, HMSET, HMGET
 
 ## v0.9.0 (2022-07-31)
 - Initial release